requirements.txt
Flask == 3.0.0
requests == 2.31.0
pymongo == 4.6.0
PyYAML == 6.0.1
Flask-Cors == 4.0.0
bcrypt == 4.0.1
pandas